+++ src/App.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 
 import './app.less';
-import {SykefraværprosentProvider} from './SykefraværprosentProvider';
+import {SammenligningProvider} from './SammenligningProvider';
 import Banner from './Banner/Banner';
 import Forside from './Forside/Forside';
 
 const App: React.FC = () => {
     return (
         <div className="app">
-            <SykefraværprosentProvider>
+            <SammenligningProvider>
                 <Banner tekst="Sykefraværsstatistikk" />
                 <Forside/>
-            </SykefraværprosentProvider>
+            </SammenligningProvider>
         </div>
     );
 };
