--- src/mocking/sykefravarsprosent.json
@@ -1,4 +1,0 @@
-{
-  "land": 6.5,
-  "kvartal": 2
-}
